Remove Int32 value representation, handle unsigned values better
A integer value in a FIT file can be signed/unsigned or different length
and use a special value in its range to denote an illegal value. I
suspect the handling of these details were not correct so far and this
is trying to simplify it. Always use int values on the OCaml side to
represent a value and try to handle unsinged values better to ensure the
repesentation in OCaml is correct.

Removing the Int32 representation is an interface change that is
potentially breaking for clients.

Signed-off-by: Christian Lindig <christian.lindig@citrix.com>